# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduG School Policy API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "administratorBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "deviceCredential": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "DeviceCredential followed by the enrolled credential UUID. Requests also carry X-Device-Timestamp, X-Device-Nonce, X-Device-Body-SHA256 and X-Device-Signature."
      }
    },
...
  },
  "security": [
    {
      "administratorBearer": []
    }
  ],
  "paths": {
...
    "/public/translation/translate": {
      "post": {
        "summary": "Translate approved public marketing content",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["content_key", "target_language"],
                "properties": {
                  "content_key": {"type": "string", "pattern": "^(header|hero|capabilities|how|context|status)\\."},
                  "target_language": {"type": "string", "enum": ["eng", "ach", "lgg", "teo", "nyn", "lug"]}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Translated marketing content or canonical English fallback"},
          "400": {"description": "Invalid or unapproved translation request"},
          "429": {"description": "Translation rate limit exceeded"}
        }
      }
    },
    "/admin/translation/translate": {
      "post": {
        "security": {
          "administratorBearer": []
        },
        "responses": {
          "200": {
            "description": "Approved dynamic content translated"
          },
          "400": {
            "description": "Invalid or unapproved translation request"
          },
          "401": {
            "description": "Authentication failed"
          },
          "429": {
            "description": "Translation rate limit exceeded"
          },
          "503": {
            "description": "Translation provider unavailable or not configured"
          }
        }
      }
    },
    "/admin/applications": {
      "get": {
        "responses": {
          "200": {
            "description": "Managed application catalogue"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      },
      "post": {
        "security": {
          "administratorBearer": []
        },
        "responses": {
          "201": {
            "description": "Managed application created"
          },
          "400": {
            "description": "Invalid application"
          },
          "409": {
            "description": "Application conflict"
          },
          "503": {
            "description": "Persistence unavailable"
          }
        }
      }
    },
    "/admin/applications/{application_uuid}": {
      "parameters": [
        {
          "name": "application_uuid",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "patch": {
        "security": {
          "administratorBearer": []
        },
        "responses": {
          "200": {
            "description": "Managed application updated"
          },
          "400": {
            "description": "Invalid application"
          },
          "404": {
            "description": "Application not found"
          },
          "409": {
            "description": "Application conflict"
          },
          "503": {
            "description": "Persistence unavailable"
          }
        }
      }
    },
...
    "/admin/enrollment-tokens/{token_uuid}/revoke": {
      "parameters": [
        {
          "name": "token_uuid",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Enrollment token revoked"
          }
        }
      }
    },
    "/admin/devices/{device_uuid}/credentials/revoke": {
      "parameters": [
        {
          "name": "device_uuid",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "description": "Requires device_credential.revoke permission; revokes the active credential.",
        "parameters": [
          {
            "name": "device_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device credential revoked"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "Permission denied"
          },
          "404": {
            "description": "No active credential"
          },
          "500": {
            "description": "Persistence failed"
          }
        }
      }
    },
...
    "/devices/{device_uuid}/credentials/rotate": {
      "parameters": [
        {
          "name": "device_uuid",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "security": [],
        "responses": {
          "201": {
            "description": "Credential rotated"
          }
        }
      }
    },
...
  }
}